requirements.txt
amqp==5.3.1
asgiref==3.10.0
bcrypt==5.0.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cacheops==7.2
django-cors-headers==4.9.0
django-health-check==3.20.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.16.1
dotenv==0.9.9
drf-standardized-errors==0.15.0
funcy==2.0
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pybreaker==1.4.1
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==5.2.1
requests==2.32.5
sentry-sdk==2.43.0
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise
gunicorn
django-storages
dj-database-url
uvicorn
boto3